+++ NEWS.md
@@ -3,6 +3,7 @@
 - `generate()` and `chat()` accept multiple images as prompts/messages.
 - Add functions to validate messages for `chat()` function: `validate_message()`, `validate_messages()`.
 - Add `encode_images_in_messages()` to encode images in messages for `chat()` function.
+- Add `create_messages()` to create messages easily.
 
 # ollamar 1.2.0